fix: batching
